Improve user profile and supplier search
