+++ app/routes/_auth+/onboarding.tsx
@@ -166,7 +166,9 @@ export default function OnboardingRoute() {
 		<div className="container flex min-h-full flex-col justify-center pb-32 pt-20">
 			<div className="mx-auto w-full max-w-lg">
 				<div className="flex flex-col gap-3 text-center">
-					<h1 className="text-h1">Welcome aboard {data.email}!</h1>
+					<h1 className="text-h3 md:text-h2 lg:text-h1">
+						Welcome aboard {data.email}!
+					</h1>
 					<p className="text-body-md text-muted-foreground">
 						Please enter your details.
 					</p>
