+++ prisma/schema.prisma
@@ -149,7 +149,7 @@ model Session {
 model Permission {
   id          String @id @default(cuid())
   action      String // e.g. create, read, update, delete
-  entity      String // e.g. note, user, etc.
+  entity      String // e.g. event, user, etc.
   access      String // e.g. own or any
   description String @default("")
